Remove the list tables action
https://github.com/secret-traitor/secret-traitor/issues/1

My lambda does not have the privileges for listing all Dyanmo tables in
the account. Because the list tables action was added in for debugging,
I am removing the list tables action, since that will be easier than
figuring out the syntax for adding a service-wide permission to the
handler.