Add a TLS cert
https://github.com/secret-traitor/secret-traitor/issues/4

This adds a TLS cert. The domain name has been transferred to the
account in which I intend to host this site, which means that this app
is now presuming the unusual setup of a name being registered to an
account, but a hosted zone not yet existing for the name. This app is
now also path-dependent, because I've gone in the order:
    1. Transfer name to account
    2. Create hosted zone
    3. Edit name servers for the domain to make it authoritative
    4. Create ACM cert
So this app as-is would not work in a new account. This should be fixed.