Retain the Hosted Zone on delete
https://github.com/secret-traitor/secret-traitor/issues/4

https://docs.aws.amazon.com/cdk/latest/guide/resources.html#resources_removal

The most common situation for deploying an app using this repo or a repo
based on this repo will be when you have:
    1. Set up the account where the resources will live
    2. Purchased the domain name for the app within the account
    3. Created a Hosted Zone for the domain name within the account
This is not the setup that I have currently. As I develop this stack,
I've transferred the domain name from another account into the account
where the app will ultimately live. Anyway, the Hosted Zone needs to
live outside of the stack. So I'm adding a deletion policy of RETAIN, so
that I can switch to [this
technique](https://github.com/douglasnaphas/aws-cdk-examples/blob/3017879482bb6f94e9b6629833f10a90b090b7e0/typescript/static-site/static-site.ts#L26)
of referencing an already-existing Hosted Zone in the stack. When I
switch over to that method, the the resource will be deleted from the
stack. When that happens, I don't want the physical Hosted Zone to be
destroyed.