Reduce tests to placeholders
https://github.com/secret-traitor/secret-traitor/issues/6

`make test` gives odd results--it may have special meaning. I changed
the test target name to tests.

I am putting off infra testing. My stack is environment-dependent--it
relies on the account in which it is created already having a hosted
zone with authority over a registered domain. Without that, I can't have
the cert in the stack. I am deferring testing rather than putting in the
work to figure out how to deal with this kind of environment-dependence
in tests (or remove the env dependence).