Add an example serverless.yml file
This adds
https://github.com/michalkvasnicak/aws-lambda-graphql/blob/5652b24185e7dfd08f479d14868c77a5ca62b155/packages/chat-example-server/serverless.yml
as-is to our project. The next step is to compile this file into
CloudFormation in our build, and then include the compiled template in
our CDK stack. Maybe before that we need to change the handler names to
match ours.

Raw file curled to create the file committed here:
https://raw.githubusercontent.com/michalkvasnicak/aws-lambda-graphql/5652b24185e7dfd08f479d14868c77a5ca62b155/packages/chat-example-server/serverless.yml