Install some deps that sls CLI complained about
We may have a problem here: even after this commit, `npx serverless
package` complains about packages referenced _in backend_ that are
missing. I think `npx serverless package` (which we need to do in order
to get a CloudFormation template that we can include in our CDK stack)
tries to go and package up whatever is referenced in backend.