Don't set up WebSiteBucket for static hosting
The rationale is explained in [this and the following
comment](https://github.com/secret-traitor/secret-traitor/commit/ed93d376333f36149a65fdb9908e174a1d31b353#r45460750).

I wonder if the strange result we're seeing (gh-29) with the website
bucket not existing might be because I have a policy in place within
this account's organization that prevents buckets from having public
read access. That policy could be in place thanks to the AWS service
Control Tower. As explained in the comment linked above, we do not need
public read access for our frontend bucket.